# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.OrderItemPack:
    properties:
      order_item:
        items:
          $ref: '#/definitions/model.OrderItemModel'
        type: array
      table_id:
        type: string
    type: object
...
  model.OrderItemModel:
    properties:
      created_at:
        type: string
      food_id:
        type: string
      id:
        type: string
      order_id:
        type: string
      order_item_id:
        type: string
      quantity:
        type: string
      unit_price:
        type: number
      updated_at:
        type: string
    required:
    - food_id
    - order_id
    - quantity
    - unit_price
    type: object
...
host: localhost:3000
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive Restaurant Management System API for managing users,
    menus, food items, orders, tables, invoices, and notes. Available on both local
    and online servers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
  x-servers:
  - description: Local Development Server
    url: http://localhost:3000/api/v1
  - description: Online Production Server
    url: https://restaurant-management-f9kx.onrender.com/api/v1
paths:
...
  /orderItem:
    post:
      consumes:
      - application/json
      description: Create one or more order items associated with an order
      parameters:
      - description: Order items data with table_id and items list
        in: body
        name: orderItemPack
        required: true
        schema:
          $ref: '#/definitions/controller.OrderItemPack'
      produces:
      - application/json
      responses:
        "200":
          description: Order items created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create order items for an order
      tags:
      - OrderItem
  /orderItem/{id}:
    get:
      description: Retrieve all items for a specific order with details (food, table,
        etc)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of order items with details
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get order items by order ID
      tags:
      - OrderItem
    put:
      consumes:
      - application/json
      description: Update order item details by order item ID
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated order item data
        in: body
        name: orderItem
        required: true
        schema:
          $ref: '#/definitions/model.OrderItemModel'
      produces:
      - application/json
      responses:
        "200":
          description: Order item updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an order item
      tags:
      - OrderItem
  /orderItems:
    get:
      description: Retrieve a list of all order items across all orders
      produces:
      - application/json
      responses:
        "200":
          description: List of order items
          schema:
            items:
              $ref: '#/definitions/model.OrderItemModel'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all order items
      tags:
      - OrderItem
...
schemes:
- http
- https
swagger: "2.0"